refactor: share the daemon client and the reprime envelope
The fleet tool server has to reach the daemon over hangar.sock and has to
fence agent-authored text the same way the re-prime corpus does. Both
already existed, one inside the CLI and one inside ainb-acp, so they move
to crates every caller can reach rather than being written a third time.
The old paths re-export, so nothing that used them changes.
