+++ ainb-tui/crates/ainb-core/Cargo.toml
@@ -126,6 +126,9 @@ ainb-hangar-core = { path = "../ainb-hangar-core", version = "0.1.0" }
 # Pure wire types for the daemon socket — the bridge reads `attention/list` and
 # routes answers through `attention/answer` on the control plane (D18).
 ainb-hangar-proto = { path = "../ainb-hangar-proto", version = "0.1.0" }
+# The one socket client (auth/hello + Content-Length JSON-RPC + the retained
+# subscriptions), re-exported unchanged at `crate::fleet::bridge::daemon`.
+ainb-hangar-client = { path = "../ainb-hangar-client", version = "0.1.0" }
 ainb-hangar-store = { path = "../ainb-hangar-store", version = "0.1.0" }
 ainb-hangar-daemon = { path = "../ainb-hangar-daemon" }
 # sqlx — the hangar CLI workspace-bootstrap path issues raw SQL against the
