+++ ainb-tui/crates/ainb-hangar-proto/src/lib.rs
@@ -20,6 +20,7 @@ pub mod fleet;
 pub mod lifecycle;
 pub mod methods;
 pub mod pr_status;
+pub mod reprime;
 pub mod settings;
 pub mod snapshots;
